+++ semcore/radio/__tests__/radio.axe-test.tsx
@@ -1,0 +1,26 @@
+import { expect, test, getAccessibilityViolations } from '@semcore/testing-utils/playwright';
+import { e2eStandToHtml } from '@semcore/testing-utils/e2e-stand';
+
+test.describe('Radio', () => {
+  test('Radiogroup', async ({ page }) => {
+    const standPath = 'stories/components/radio/docs/examples/radiogroup_example.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+    await page.setContent(htmlContent);
+
+    const violations = await getAccessibilityViolations({ page });
+
+    expect(violations).toEqual([]);
+  });
+
+  test('Additional props to input', async ({ page }) => {
+    const standPath = 'stories/components/radio/docs/examples/additional_props_for_input.tsx';
+    const htmlContent = await e2eStandToHtml(standPath, 'en');
+
+    await page.setContent(htmlContent);
+
+    const violations = await getAccessibilityViolations({ page });
+
+    expect(violations).toEqual([]);
+  });
+});