+++ stories/components/spin/tests/examples/spin-centered.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import Spin from '@semcore/spin';
+
+const Demo = () => {
+  return (
+    <>
+    <div style={{ display: 'flex', width: 200, height: 200 }}>
+        <Spin centered />
+      </div>
+    </>
+  );
+};
+
+export default Demo;