+++ semcore/card/CHANGELOG.md
@@ -8,6 +8,18 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 - Major version.
 
+## [5.48.2] - 2025-05-13
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.48.4 ~> 4.48.5], `@semcore/button` [5.43.2 ~> 5.43.3]).
+
+## [5.48.1] - 2025-05-09
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/utils` [4.48.2 ~> 4.48.4], `@semcore/button` [5.43.1 ~> 5.43.2], `@semcore/icon` [4.61.0 ~> 4.62.0]).
+
 ## [5.48.0] - 2025-04-11
 
 ### Changed