+++ semcore/i18n-unplugin/CHANGELOG.md
@@ -8,6 +8,18 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 - Major version.
 
+## [0.18.22] - 2025-05-13
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/core` [2.39.3 ~> 2.39.4]).
+
+## [0.18.21] - 2025-05-09
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/core` [2.39.2 ~> 2.39.3]).
+
 ## [0.18.20] - 2025-04-04
 
 ### Changed