+++ semcore/notice-global/CHANGELOG.md
@@ -8,6 +8,18 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 
 - Major version.
 
+## [2.62.2] - 2025-05-13
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/button` [5.43.2 ~> 5.43.3], `@semcore/utils` [4.48.4 ~> 4.48.5], `@semcore/icon` [4.62.0 ~> 4.62.1], `@semcore/flex-box` [5.41.3 ~> 5.41.4], `@semcore/animation` [2.41.3 ~> 2.41.4], `@semcore/core` [2.39.3 ~> 2.39.4]).
+
+## [2.62.1] - 2025-05-09
+
+### Changed
+
+- Version patch update due to children dependencies update (`@semcore/button` [5.43.1 ~> 5.43.2], `@semcore/utils` [4.48.2 ~> 4.48.4], `@semcore/icon` [4.61.0 ~> 4.62.0], `@semcore/flex-box` [5.41.2 ~> 5.41.3], `@semcore/animation` [2.41.2 ~> 2.41.3], `@semcore/core` [2.39.2 ~> 2.39.3]).
+
 ## [2.62.0] - 2025-04-11
 
 ### Changed